refactor: new card page use hooks
