+++ src/pages/createCardPage/createCardPage.test.tsx
@@ -1,6 +1,6 @@
 import { describe, test, expect } from 'vitest';
 import { render } from '@testing-library/react';
-import CreateCardPage from './createCardPage';
+import CreateCardPage from './newCardPage';
 
 describe('<CreateCardPage />', () => {
   test('CreateCardPage mounts properly', () => {
